Move generic Thomas-solver scans to common/math/tridiagonal
The forward-sweep and back-substitution scan operators are a generic
TDMA solver; both tmx vertical diffusion solvers now import them from
common. The dycore w solver keeps its fused physics-specific variant.
