+++ model/atmosphere/subgrid_scale_physics/tmx/src/icon4py/model/atmosphere/subgrid_scale_physics/tmx/stencils/solve_vertical_diffusion_cells.py
@@ -8,48 +8,11 @@
 import gt4py.next as gtx
 
 from icon4py.model.common import dimension as dims, field_type_aliases as fa
+from icon4py.model.common.math.tridiagonal import (
+    _tridiagonal_back_substitution,
+    _tridiagonal_forward_sweep,
+)
 from icon4py.model.common.type_alias import wpfloat
-
-
-@gtx.scan_operator(axis=dims.KDim, forward=True, init=(wpfloat("0.0"), wpfloat("0.0")))
-def _tridiagonal_forward_sweep(
-    state_kminus1: tuple[wpfloat, wpfloat],
-    a: wpfloat,
-    b: wpfloat,
-    c: wpfloat,
-    d: wpfloat,
-) -> tuple[wpfloat, wpfloat]:
-    """
-    Forward elimination of the Thomas algorithm ('tdma_solver_vec',
-    iconmath mo_math_utilities.F90):
-        m       = 1 / (b(k) - c_prime(k-1) * a(k))
-        c_prime = c(k) * m
-        d_prime = (d(k) - d_prime(k-1) * a(k)) * m
-
-    The init state (0, 0) reproduces the Fortran first-row formulas
-    c_prime = c/b and d_prime = d/b for any value of a at the first row.
-    """
-    c_prime_kminus1 = state_kminus1[0]
-    d_prime_kminus1 = state_kminus1[1]
-    m = wpfloat("1.0") / (b - c_prime_kminus1 * a)
-    c_prime = c * m
-    d_prime = (d - d_prime_kminus1 * a) * m
-    return c_prime, d_prime
-
-
-@gtx.scan_operator(axis=dims.KDim, forward=False, init=wpfloat("0.0"))
-def _tridiagonal_back_substitution(
-    x_kplus1: wpfloat,
-    c_prime: wpfloat,
-    d_prime: wpfloat,
-) -> wpfloat:
-    """
-    Back substitution of the Thomas algorithm:
-        x(k) = d_prime(k) - c_prime(k) * x(k+1)
-
-    The init state 0 reproduces the Fortran last-row formula x = d_prime.
-    """
-    return d_prime - c_prime * x_kplus1
 
 
 @gtx.field_operator
